+++ bot.py
@@ -182,11 +182,13 @@ class ModmailBot(commands.Bot):
         else:
             await self.threads.populate_cache()
 
-
     async def process_modmail(self, message):
         """Processes messages sent to the bot."""
 
-        reaction = '🚫' if message.author.id in self.blocked_users else '✅'
+        blocked_emoji = self.config.get('blocked_emoji', '🚫')
+        sent_emoji = self.config.get('sent_emoji', '✅')
+
+        reaction = blocked_emoji if message.author.id in self.blocked_users else sent_emoji
 
         try:
             await message.add_reaction(reaction)